+++ app/Models/ApiToken.php
@@ -2,12 +2,16 @@
 
 namespace App\Models;
 
+use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\Relations\HasMany;
 use Laravel\Sanctum\PersonalAccessToken;
 
 class ApiToken extends PersonalAccessToken
 {
+    /** @use HasFactory<\Database\Factories\ApiTokenFactory> */
+    use HasFactory;
+
     protected $table = 'personal_access_tokens';
 
     protected $fillable = ['name', 'description', 'token', 'abilities', 'expires_at', 'revoked_at', 'created_by'];